+++ helpers/hdocker.py
@@ -1009,107 +1009,6 @@ def run_dockerized_pandoc(
         ret = None
     return ret
 
-# #############################################################################
-# Dockerized extract_notebook_images.
-# #############################################################################
-
-def run_dockerized_notebook_image_extractor(
-    notebook_path: str,
-    output_dir: str,
-    *,
-    force_rebuild: bool = False,
-    use_sudo: bool = False,
-) -> None:
-    """
-    Run NotebookImageExtractor in a Docker container.
-
-    """
-    _LOG.debug(hprint.func_signature_to_str())
-    # Build the container image, if needed.
-    container_image = "tmp.notebook_image_extractor"
-    dockerfile = r"""
-    # TODO(gp): This might be problematic on MacOS / ARM.
-    FROM --platform=linux/amd64 python:3.10-slim
-
-    # Install required system libraries for Chromium and Playwright.
-    RUN apt-get update && apt-get install -y \
-        libglib2.0-0 \
-        libnss3 \
-        libnspr4 \
-        libdbus-1-3 \
-        libatk1.0-0 \
-        libatk-bridge2.0-0 \
-        libexpat1 \
-        libatspi2.0-0 \
-        libdbus-glib-1-2 \
-        libxcomposite1 \
-        libxdamage1 \
-        libxfixes3 \
-        libxrandr2 \
-        libgbm1 \
-        libxkbcommon0 \
-        libasound2 \
-        libcups2 \
-        libpango-1.0-0 \
-        libcairo2 \
-        && rm -rf /var/lib/apt/lists/*
-
-    # Set the environment variable for Playwright to install browsers in a known location.
-    ENV PLAYWRIGHT_BROWSERS_PATH=/ms-playwright
-
-    # Create the directory for Playwright browsers and ensure it's writable.
-    RUN mkdir -p /ms-playwright && chmod -R 777 /ms-playwright
-
-    RUN pip install nbconvert nbformat playwright pyyaml
-
-    RUN python -m playwright install
-
-    WORKDIR /app
-    """
-    container_image = build_container_image(
-        container_image, dockerfile, force_rebuild, use_sudo
-    )
-    # Convert file paths to Docker paths.
-    is_caller_host = not hserver.is_inside_docker()
-    use_sibling_container_for_callee = True
-    caller_mount_path, callee_mount_path, mount = get_docker_mount_info(
-        is_caller_host, use_sibling_container_for_callee
-    )
-    notebook_path = convert_caller_to_callee_docker_path(
-        notebook_path,
-        caller_mount_path,
-        callee_mount_path,
-        check_if_exists=True,
-        is_input=True,
-        is_caller_host=is_caller_host,
-        use_sibling_container_for_callee=use_sibling_container_for_callee,
-    )
-    output_dir = convert_caller_to_callee_docker_path(
-        output_dir,
-        caller_mount_path,
-        callee_mount_path,
-        check_if_exists=False,
-        is_input=False,
-        is_caller_host=is_caller_host,
-        use_sibling_container_for_callee=use_sibling_container_for_callee,
-    )
-    executable = get_docker_executable(use_sudo)
-    python_code = (
-        "from helpers.hjupyter import NotebookImageExtractor; "
-        "extractor = NotebookImageExtractor({}, {}); "
-        "extractor._extract_and_capture()"
-    ).format(json.dumps(notebook_path), json.dumps(output_dir))
-    inner_cmd = f"python -c {json.dumps(python_code)}"
-    docker_cmd = (
-        f"{executable} run --rm --user $(id -u):$(id -g) "
-        f"-e PLAYWRIGHT_BROWSERS_PATH=/ms-playwright "
-        f"--workdir {callee_mount_path} --mount {mount} "
-        f"{container_image} "
-        f'bash -c {json.dumps(inner_cmd)}'
-    )
-
-    hsystem.system(docker_cmd)
-    hsystem.system(docker_cmd)
 
 # #############################################################################
 # Dockerized markdown_toc.
@@ -1553,107 +1452,6 @@ def dockerized_tikz_to_bitmap(
         force_rebuild=force_rebuild,
         use_sudo=use_sudo,
     )
-
-
-# #############################################################################
-# Dockerized llm_transform.
-# #############################################################################
-
-
-def run_dockerized_llm_transform(
-    in_file_path: str,
-    cmd_opts: List[str],
-    out_file_path: str,
-    *,
-    return_cmd: bool = False,
-    force_rebuild: bool = False,
-    use_sudo: bool = False,
-) -> Optional[str]:
-    """
-    Run dockerized_llm_transform.py in a Docker container with all its dependencies.
-    """
-    _LOG.debug(hprint.func_signature_to_str())
-    #
-    hdbg.dassert_in("OPENAI_API_KEY", os.environ)
-    hdbg.dassert_isinstance(cmd_opts, list)
-    # Build the container, if needed.
-    container_image = "tmp.llm_transform"
-    dockerfile = r"""
-    FROM python:3.12-alpine
-
-    # Install Bash.
-    #RUN apk add --no-cache bash
-
-    # Set Bash as the default shell.
-    #SHELL ["/bin/bash", "-c"]
-
-    # Install pip packages.
-    RUN pip install --no-cache-dir openai
-    """
-    container_image = build_container_image(
-        container_image, dockerfile, force_rebuild, use_sudo
-    )
-    # Convert files to Docker paths.
-    is_caller_host = not hserver.is_inside_docker()
-    use_sibling_container_for_callee = True
-    caller_mount_path, callee_mount_path, mount = get_docker_mount_info(
-        is_caller_host, use_sibling_container_for_callee
-    )
-    in_file_path = convert_caller_to_callee_docker_path(
-        in_file_path,
-        caller_mount_path,
-        callee_mount_path,
-        check_if_exists=True,
-        is_input=True,
-        is_caller_host=is_caller_host,
-        use_sibling_container_for_callee=use_sibling_container_for_callee,
-    )
-    out_file_path = convert_caller_to_callee_docker_path(
-        out_file_path,
-        caller_mount_path,
-        callee_mount_path,
-        check_if_exists=False,
-        is_input=False,
-        is_caller_host=is_caller_host,
-        use_sibling_container_for_callee=use_sibling_container_for_callee,
-    )
-    helpers_root = hgit.find_helpers_root()
-    helpers_root = convert_caller_to_callee_docker_path(
-        helpers_root,
-        caller_mount_path,
-        callee_mount_path,
-        check_if_exists=True,
-        is_input=False,
-        is_caller_host=is_caller_host,
-        use_sibling_container_for_callee=use_sibling_container_for_callee,
-    )
-    git_root = hgit.find_git_root()
-    script = hsystem.find_file_in_repo("dockerized_llm_transform.py", root_dir=git_root)
-    script = convert_caller_to_callee_docker_path(
-        script,
-        caller_mount_path,
-        callee_mount_path,
-        check_if_exists=True,
-        is_input=True,
-        is_caller_host=is_caller_host,
-        use_sibling_container_for_callee=use_sibling_container_for_callee,
-    )
-    cmd_opts_as_str = " ".join(cmd_opts)
-    executable = get_docker_executable(use_sudo)
-    docker_cmd = (
-        f"{executable} run --rm --user $(id -u):$(id -g)"
-        f" -e OPENAI_API_KEY -e PYTHONPATH={helpers_root}"
-        f" --workdir {callee_mount_path} --mount {mount}"
-        f" {container_image}"
-        f" {script} -i {in_file_path} -o {out_file_path} {cmd_opts_as_str}"
-    )
-    if return_cmd:
-        ret = docker_cmd
-    else:
-        # TODO(gp): Note that `suppress_output=False` seems to hang the call.
-        hsystem.system(docker_cmd, suppress_output=False)
-        ret = None
-    return ret
 
 
 # #############################################################################
